backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "qiskit-iqm==17.8",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.118.0",
    "matplotlib>=3.10.6",
    "notebook>=7.4.7",
    "pylatexenc>=2.10",
    "qiskit-iqm>=17.8",
]
